requirements.txt
ds_ctcdecoder==0.6.1
mako==1.0.6
matplotlib==2.0.0
numpy==1.12.0
pandas==1.1.5
protobuf==3.2.0
pyxdg==0.26
pyyaml==5.1
requests==2.20.0
scipy==0.18.1
tensorflow-gpu==1.15
tree==0.1.0q
